Move CMS-specific utils and types into starter kit as well
Prevents type errors from dependencies that aren't installed.
