+++ .ddev/commands/web/setup-drupal
@@ -6,8 +6,8 @@
 
 if [ -e ./codegen.ts ] || [ -d ./app/api ]; then
   echo "Drupal setup can only be run once."
-  echo "If you need the starter files, copy them from the starterkits directory."
-  echo 0;
+  echo "If you need the starter files, copy them from the /starterkits directory."
+  exit 0;
 fi
 
 node lib/drupal/merge-package-json.js
@@ -15,7 +15,8 @@ npm install --package-lock-only
 npm ci
 cp ./starterkits/drupal/codegen.ts ./codegen.ts
 cat ./starterkits/drupal/.env.drupal >> .env.local.example
-cp -r ./starterkits/drupal/api ./app
 cp -r ./starterkits/drupal/app/* ./app
+cp -r ./starterkits/drupal/util/* ./util
+cp -r ./starterkits/drupal/types/* ./types
 pm2 restart frontend
 pm2 restart storybook
