+++ lib/drupal/generate-graphql-types.ts
@@ -1,12 +1,18 @@
+// The GraphQL codegen module is only installed on headless projects using
+// GraphQL (namely, Drupal), which is also the only type of project this script
+// applies to.
+// @ts-ignore
 import { CodegenContext, generate } from '@graphql-codegen/cli';
 import { resolve } from 'node:path';
-// The codegen file won't be at its location until setup-drupal is run. Ignore
-// the missing file, because @ts-expect-error then errors once setup-drupal is
-// run and the file *is* there.
+// The codegen and util files won't be at their destination until the setup
+// script is run. Using ts-ignore here because ts-expect-error will then error
+// once the script is run and the files *do* exist.
 // @ts-ignore
-import codegenConfig from '../../codegen';
-import drupal from '../../util/drupal/drupal';
-import graphqlEndpoint from '../../util/drupal/graphqlEndpoint';
+import codegenConfig from '@/codegen';
+// @ts-ignore
+import drupal from '@/util/drupal';
+// @ts-ignore
+import graphqlEndpoint from '@/util/drupal/graphqlEndpoint';
 
 async function generateGraphQLTypes() {
   try {
