+++ starterkits/drupal/util/dataIsEntityType.ts
