+++ starterkits/drupal/util/graphqlEndpoint.ts
