+++ tsconfig.json
@@ -44,6 +44,6 @@
   "exclude": [
     "node_modules",
     "lib/plop-templates/*",
-    "starterkits/*/examples/**"
+    "starterkits"
   ]
 }
